release: v0.3.0 prep — version bumps, site judge section, clippy fix
- Cargo workspace + npm wrapper + optionalDependencies pins -> 0.3.0
  (version-sanity gate replicated locally, all three checks pass)
- site: Semantic judge section + TOC entry (mobile nav derives from TOC)
- tests/judge.rs: clippy unnecessary-cast fix (ci job gates publish with -D warnings)

+++ crates/core/tests/judge.rs
@@ -238,7 +238,7 @@ fn adversarial_answers_never_move_the_deterministic_score() {
         let findings: Vec<Value> = (0..n)
             .map(|k| {
                 json!({
-                    "lines": [1 + (k as u64), 2 + (k as u64)],
+                    "lines": [1 + k, 2 + k],
                     "quote": format!("q{seed:x}{k}"),
                     "rationale": format!("r{seed:x}"),
                     "confidence": conf,
